Fallback for user object in vote event
Noticed some undefined user objects, but the ID should still be provided, so a fallback to allow most miniplug methods would be very useful.

The front end on plug.dj seems to not emit this event to the API (but still takes the vote into account internally). I didn't add that but it might be something worth considering?